+++ common/src/main/java/io/github/stainlessstasis/config/manager/AbstractConfigManager.java
@@ -1,6 +1,7 @@
 package io.github.stainlessstasis.config.manager;
 
 import com.google.gson.*;
+import io.github.stainlessstasis.alert.Tag;
 import io.github.stainlessstasis.config.client.PokemonConfig;
 import io.github.stainlessstasis.core.CobblemonSpawnAlerts;
 import io.github.stainlessstasis.platform.Services;
@@ -160,6 +161,7 @@ public abstract class AbstractConfigManager {
 
                     mergeJsonObjects(specificPokemonDefault, userSpecificConfig);
                     fixHexColor(specificPokemonDefault);
+                    fixOutdatedNaming(specificPokemonDefault);
                     mergedPokemonConfigs.add(pokemonName, specificPokemonDefault);
                 } else {
                     CobblemonSpawnAlerts.LOGGER.warn("Invalid entry for Pokemon '"+pokemonName+"' in config file `"+fileName+"`. Skipping.");
@@ -182,6 +184,22 @@ public abstract class AbstractConfigManager {
             String hex = configObject.get("glowColor").getAsString();
             if (!hex.isEmpty() && !hex.startsWith("#")) {
                 configObject.addProperty("glowColor", "#" + hex);
+            }
+        }
+    }
+
+    private void fixOutdatedNaming(JsonObject configObject) {
+        if (configObject.has("statDisplayModes") && configObject.get("statDisplayModes").isJsonObject()) {
+            JsonObject modes = configObject.getAsJsonObject("statDisplayModes");
+
+            if (modes.has("nearestPlayer")) {
+                modes.add(Tag.NEAREST_PLAYER.getKey(), modes.get("nearestPlayer"));
+                modes.remove("nearestPlayer");
+            }
+
+            if (modes.has("coordinates")) {
+                modes.add(Tag.COORDINATES.getKey(), modes.get("coordinates"));
+                modes.remove("coordinates");
             }
         }
     }